prereq-scraper: merge into sis-scraper
This is the end of the line for the prerequisties scraper as a standalone
scraper. Since we retrieve the same data now for both, it only made
sense to merge them in to reduce duplicate work (and the load on SIS).

This change also cleans up scrape.yml to remove the .env
generation and the discrete scraping step.
